+++ idf_analysis/in_out.py
@@ -1,6 +1,7 @@
+from collections import OrderedDict
+
 import pandas as pd
 import yaml
-from collections import OrderedDict
 
 
 def import_series(filename, series_label='precipitation', index_label='datetime', csv_reader_args=None):
@@ -19,7 +20,7 @@ def import_series(filename, series_label='precipitation', index_label='datetime'
         if csv_reader_args is None:
             csv_reader_args = dict(sep=';', decimal=',')
         try:
-            ts = pd.read_csv(filename, index_col=0, header=0, squeeze=True, **csv_reader_args)
+            ts = pd.read_csv(filename, index_col=0, header=0, **csv_reader_args).squeeze('columns')
             ts.index = pd.to_datetime(ts.index)
             ts.index.name = index_label
             ts.name = series_label
