+++ idf_analysis/_console_script.py
@@ -5,7 +5,4 @@ def command_line_tool():
     """
     Execute the command line tool
     """
-    import matplotlib.pyplot as plt
-    plt.style.use('bmh')
-
     IntensityDurationFrequencyAnalyse.command_line_tool()
