### 2024-11-07 65eeee6
style: simplified requirements
