### 2024-11-12 4da1308
style: parquet as optional package that the user has to install
