+++ examples/example_python_api_extended.py
@@ -11,6 +11,7 @@ output_directory = Path(__file__).parent / 'ehyd_112086_2019_idf_data'
 idf = IntensityDurationFrequencyAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=True)
 
 # reading the pandas series of the precipitation (data from ehyd.gv.at - ID=112086)
+# You need to install `pyarrow` or `fastparquet` to read and write parquet files.
 series = pd.read_parquet('ehyd_112086.parquet')['N-Minutensummen-112086']
 
 # setting the series for the analysis
