+++ examples/example_python_api.py
@@ -13,6 +13,7 @@ output_directory = Path('ehyd_112086_idf_data')
 idf = HeavyRainfallIndexAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=True)
 
 # reading the pandas series of the precipitation (data from ehyd.gv.at - ID=112086)
+# You need to install `pyarrow` or `fastparquet` to read and write parquet files.
 series = pd.read_parquet('ehyd_112086.parquet').squeeze()
 series.name = 'precipitation'
 
