+++ examples/convert_old_to_new_parameters.py
@@ -9,6 +9,7 @@ output_directory = path.join('ehyd_112086_idf_data')
 idf = IntensityDurationFrequencyAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=True)
 
 # reading the pandas series of the precipitation (data from ehyd.gv.at - ID=112086)
+# You need to install `pyarrow` or `fastparquet` to read and write parquet files.
 # series = pd.read_parquet('ehyd_112086.parquet')['precipitation']
 
 # setting the series for the analysis
