+++ examples/example_60m.py
@@ -8,6 +8,7 @@ from idf_analysis.definitions import SERIES, METHOD
 idf = IntensityDurationFrequencyAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=True)
 
 # reading the pandas series of the precipitation (data from ehyd.gv.at - ID=112086)
+# You need to install `pyarrow` or `fastparquet` to read and write parquet files.
 series = pd.read_parquet('ehyd_112086.parquet')['precipitation']
 
 series60 = series.resample('60min').sum()
