### 2024-11-12 e287186
style: tqdm as optional package that when installed, shows a progressbar and if not, doesn't show anything
