### 2024-11-12 d5fd380
style: remove statsmodels as dependency
