### 2024-11-22 #7437 @kthyng
@markuspic 👋
* How would a user test your software or otherwise verify after installation or making changes that it is running correctly?
