+++ .github/workflows/package_release.yaml
@@ -22,6 +22,10 @@ jobs:
           fetch-depth: 0
           token: ${{ secrets.GH_TOKEN }}
 
+      - name: Preprocess README.md
+        run: |
+          sed -i 's#\](\/#](https://github.com/MarkusPic/intensity_duration_frequency_analysis/blob/main/#g' README.md
+
 #      - name: pip install python-build package
 #        run: |
 #          pip install -q build
