+++ .github/workflows/sphinx_docs_to_gh_pages.yaml
@@ -42,6 +42,10 @@ jobs:
         run: |
           sudo apt-get install pandoc -y
       # ------------------------------
+      #- name: Preprocess README.md
+      #  run: |
+      #    sed -i 's#\](\/#](https://github.com/MarkusPic/intensity_duration_frequency_analysis/blob/main/#g' README.md
+
       - name: Copy readme and examples
         #            pwd
         #            ls -lh
