### 2024-11-24 #7437 @MarkusPic
Users can verify the correct functionality of the software by running [the example](https://github.com/MarkusPic/intensity_duration_frequency_analysis/blob/main/examples/example_python_api.ipynb) provided in the repository’s examples folder. This example is a Jupyter Notebook that demonstrates the use of the software with open rainfall monitoring data from a station in Graz, Austria, provided by the Austrian government.

The notebook includes step-by-step usage instructions, interim parameters, and plots of results, allowing users to compare their outputs with the expected outcomes. Users often test the package by reproducing this example, as it provides a clear and practical way to check whether the software is working as intended.

While this example serves as a practical testing method, I acknowledge that it does not replace proper unit testing practices. However, it has proven helpful for users to verify installation and functionality.
