### 2024-12-02 #7437 @kthyng
@MarkusPic Ok thank you. Your example notebooks would also be great additions to your docs, just a suggestion.
