### 2024-12-02 #7437 @MarkusPic
Thank you for your feedback. It actually is part of the docs: https://markuspic.github.io/intensity_duration_frequency_analysis/examples/example_python_api.html
