### 2024-12-19 #7607 @editorialbot
[REVIEW]: idf-analysis: Heavy rainfall intensity as a function of duration and return period
**Submitting author:** @MarkusPic (<a href="http://orcid.org/0000-0002-8101-2163">Markus Pichler</a>)
**Repository:** https://github.com/MarkusPic/intensity_duration_frequency_analysis
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.1
