### 2025-01-03 #7607 @einaraz
Dear @MarkusPic,

I have started reviewing idf-analysis and have some initial questions, which I’ve detailed below. Since I'm less familiar with this type of analysis, my evaluation primarily focuses on the software and its reproducibility. I apologize in advance if any of my questions or comments seem unclear.

**Comments on the paper**

Figure 1: Please define the x and y labels, as well as the legend, in the figure caption.

Lines 37–39: I don't fully understand the sentence “For medium-length time series, a partial series analysis employs the exponential distribution, using the largest 2.7 times the number of measurement years’ events”. Could you please clarify or rephrase this for better comprehension?

Equation 1: I believe the duration, D, is not defined in the text. Additionally, could you please provide more detail on the formulation of u(D) and w(D)? Are there alternative options available for these variables in the package?

Line 48: The paragraph starting on this line references previous tools that address similar problems. However, it’s not clear how the current paper contributes to this area and what builds on prior work. I recommend restructuring this section to clearly differentiate the contributions of the current package from existing tools.

**Comments on the software**

I successfully installed the package without any issues. I would like to test it by working through a few examples but had some questions about the input data. Could you please add a section to the tutorial describing the required input data? Specifically, details about formatting, units, and any other relevant information would be very helpful. Since this type of analysis is new to me, I believe other users --- especially those new to Python or unfamiliar with this type of analysis --- might have similar questions.

Additionally, while reviewing the checklist, I noticed the absence of automated tests and community guidelines for seeking support or contributing. Are there plans to include these features soon?

Thank you! Please let me know if my comments are unclear.
Einara
