### 2025-01-03 336f4dd
Create FUNDING.yml
