+++ _sources/examples/example_python_api_extended.ipynb.txt
@@ -1,0 +1,675 @@
+{
+ "cells": [
+  {
+   "cell_type": "code",
+   "execution_count": 11,
+   "metadata": {
+    "ExecuteTime": {
+     "end_time": "2024-07-02T11:40:36.914661Z",
+     "start_time": "2024-07-02T11:40:36.892058Z"
+    }
+   },
+   "outputs": [],
+   "source": [
+    "from idf_analysis.idf_class import IntensityDurationFrequencyAnalyse\n",
+    "from idf_analysis.definitions import *\n",
+    "import pandas as pd\n",
+    "from os import path\n",
+    "%matplotlib inline\n",
+    "import matplotlib.pyplot as plt\n",
+    "plt.style.use('bmh')\n",
+    "\n",
+    "import matplotlib\n",
+    "matplotlib.rcParams.update({\n",
+    "    'figure.figsize': (10, 8),\n",
+    "    'figure.dpi': 100,\n",
+    "    'figure.constrained_layout.use': True,\n",
+    "})"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {
+    "collapsed": false,
+    "jupyter": {
+     "outputs_hidden": false
+    }
+   },
+   "source": [
+    "# Intensity Duration Frequency Analyse - EXTENDED"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 12,
+   "metadata": {
+    "collapsed": false,
+    "jupyter": {
+     "outputs_hidden": false
+    },
+    "pycharm": {
+     "name": "#%%\n"
+    }
+   },
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "_balance_parameter_change\n"
+     ]
+    }
+   ],
+   "source": [
+    "# sub-folder for the results\n",
+    "output_directory = path.join('ehyd_112086_idf_data')\n",
+    "\n",
+    "# initialize of the analysis class\n",
+    "idf = IntensityDurationFrequencyAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=True)\n",
+    "\n",
+    "# reading the pandas series of the precipitation (data from ehyd.gv.at - ID=112086)\n",
+    "# You need to install `pyarrow` or `fastparquet` to read and write parquet files.\n",
+    "series = pd.read_parquet('ehyd_112086.parquet').squeeze('columns') # ['precipitation']\n",
+    "\n",
+    "# setting the series for the analysis\n",
+    "idf.set_series(series)\n",
+    "\n",
+    "# auto-save the calculated parameter so save time for a later use\n",
+    "idf.auto_save_parameters(path.join(output_directory, 'idf_parameters.yaml'))"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 13,
+   "metadata": {
+    "collapsed": false,
+    "jupyter": {
+     "outputs_hidden": false
+    },
+    "pycharm": {
+     "name": "#%%\n"
+    }
+   },
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "('Columns: ',\n",
+       " Index(['start', 'end', 'duration', 'rain_sum', 'last_event'], dtype='object'),\n",
+       " '| Number of events: ',\n",
+       " 1776)"
+      ]
+     },
+     "execution_count": 13,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
+   "source": [
+    "events = idf.rain_events\n",
[…]
