### 2025-01-06 3c13b8c
fix: better caption text for event plot
