### 2025-01-06 7575811
style: new horizontal bar plot function used in IDF and SRI event plots
