+++ idf_analysis/heavy_rainfall_index.py
@@ -15,6 +15,7 @@ import pandas as pd
 from .definitions import COL
 from .idf_class import IntensityDurationFrequencyAnalyse
 from .little_helpers import duration_steps_readable, minutes_readable, frame_looper, event_caption
+from .plot_helpers import _bar_axes, RETURN_PERIOD_COLORS, _set_xlim
 from .sww_utils import guess_freq, rain_events, event_duration, resample_rain_series, rain_bar_plot
 
 # from matplotlib.colors import ListedColormap, Normalize
@@ -180,20 +181,20 @@ class HeavyRainfallIndexAnalyse(IntensityDurationFrequencyAnalyse):
             int | float | list | numpy.ndarray | pandas.Series: heavy rain index
         """
         tn = self.get_return_period(height_of_rainfall, duration)
+        # ----
 
         if self.method == self.METHODS.MUDERSBACH:
             if isinstance(tn, (pd.Series, np.ndarray)):
                 sri = np.round(1.5 * np.log(tn) + 0.4 * np.log(duration), 0)
                 sri[tn <= 1] = 1
                 sri[tn >= 100] = 12
-                return sri
             else:
                 if tn <= 1:
-                    return 1
+                    sri = 1
                 elif tn >= 100:
-                    return 12
+                    sri = 12
                 else:
-                    return np.round(1.5 * np.log(tn) + 0.4 * np.log(duration), 0)
+                    sri = np.round(1.5 * np.log(tn) + 0.4 * np.log(duration), 0)
 
         elif self.method == self.METHODS.SCHMITT:
             if isinstance(tn, (pd.Series, np.ndarray)):
@@ -226,31 +227,38 @@ class HeavyRainfallIndexAnalyse(IntensityDurationFrequencyAnalyse):
                 sri[tn < 0.5] = 0
             else:
                 if tn < 0.5:
-                    return 0
-            return np.clip(np.ceil(sri), 0, 12)
+                    sri = 0
+            sri = np.clip(np.ceil(sri), 0, 12)
 
         elif self.method == self.METHODS.SCHMITT2015:
             # TODO: additional_files.schmitt_ortsunabhaengig_2015.ods
             if duration == 15:
                 if height_of_rainfall < 10:
-                    return 0
+                    sri = 0
             elif duration == 60:
                 if height_of_rainfall < 15:
-                    return 0
+                    sri = 0
             elif duration == 120:
                 if height_of_rainfall < 20:
-                    return 0
+                    sri = 0
             elif duration == 240:
                 if height_of_rainfall < 20:
-                    return 0
+                    sri = 0
             elif duration == 360:
                 if height_of_rainfall < 25:
-                    return 0
+                    sri = 0
             else:
-                return np.nan
+                sri = np.nan
 
         else:
             raise NotImplementedError(f'Method {self.method} not implemented!')
+
+        if isinstance(tn, (pd.Series, np.ndarray)):
+            sri[tn <= 0.3] = 0
+        else:
+            if tn < 0.3:
+                sri = 0
+
         return sri
 
     # __________________________________________________________________________________________________________________
@@ -470,72 +478,20 @@ class HeavyRainfallIndexAnalyse(IntensityDurationFrequencyAnalyse):
             sri[dur] = self.get_sri(h, dur)
         return pd.Series(sri, name=self.method)
 
-    def sri_bar_axes(self, ax, sri_frame, durations=None):
-        """ create a bar axes for the sri event plot
+    def sri_bar_axes(self, ax, sri_frame):
+        """
[…]
