### 2025-01-06 3a4e242
fix: added SRI table event plot
