### 2025-01-07 #7607 @RBhupi
@MarkusPic Suggestion: Add Unit Tests for Future Releases.

I appreciate the automated GitHub workflows for documentation and pip releases in this repository. I suggest adding unit tests. Unit tests will help catch bugs early and ensure that new changes do not break existing functionality.
