### 2025-01-07 0b365c8
Update README.md
JOSS paper review statur
