### 2025-01-12 #7607 @MarkusPic
@RBhupi and @einaraz thank you both for the good and nice feedback. I am currently busy in my job due to some deadlines, but I will soon update the paper and will start adding unit tests.

The reason why there are no unit tests was because I used the example notebook with the used example data as functional test. (as I recently explained here <https://github.com/openjournals/joss-reviews/issues/7437#issuecomment-2496211766>)
