### 2025-01-29 #7607 @MarkusPic
I'm sorry that I haven't implemented the feedback yet, but unfortunately, I still have a lot of work to do. I've already implemented some of the comments on the paper, but I wanted to implement everything before uploading the changes. I've also already started writing the unit tests. Likewise, I wanted to wait until I've written at least one or two tests for all the functions.
