+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-
+<!--next-version-placeholder-->
 
 ## v0.3.0 (2024-11-06)
 
@@ -41,16 +41,6 @@
 
 * style: inline loop to multiline loop ([`9c2a92c`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/9c2a92c8d3d050f89a59ab79045e001bd5e5cd52))
 
-### Unknown
-
-* 0.2.9
-
-Automatically generated by python-semantic-release
-
-[skip actions] ([`54ee5eb`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/54ee5ebdf74185d69446caf1753e1e360a787dda))
-
-* Merge remote-tracking branch &#39;origin/main&#39; ([`cb6f64e`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/cb6f64eec4d38f72857c856a7c18ec6f51834d04))
-
 
 ## v0.2.8 (2024-08-12)
 
@@ -63,16 +53,6 @@ Automatically generated by python-semantic-release
 ### Fix
 
 * fix: limit duration steps based on series frequency and skip parameter calculation for durations ranges lower than series frequency. ([`22e9a19`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/22e9a19b7401ccbae97b98d70fa6051957481c81))
-
-### Unknown
-
-* 0.2.8
-
-Automatically generated by python-semantic-release
-
-[skip actions] ([`b5792ce`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/b5792ce1246ce84dcd51106ca7071ff392fb1301))
-
-* Merge remote-tracking branch &#39;origin/main&#39; ([`90dd636`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/90dd6365a324ad4f1463d209dc99ae9106a511c2))
 
 
 ## v0.2.7 (2024-07-22)
@@ -95,14 +75,6 @@ Automatically generated by python-semantic-release
 
 * style: adapt to new numpy version 2.0 ([`c893cef`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/c893cef7da8dce3a0fc3fd552b5257dac1cd8824))
 
-### Unknown
-
-* 0.2.7
-
-Automatically generated by python-semantic-release
-
-[skip actions] ([`13a931f`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/13a931f5f44d637bae296a20ef7a491da02c2d5a))
-
 
 ## v0.2.6 (2024-05-16)
 
@@ -113,16 +85,6 @@ Automatically generated by python-semantic-release
 ### Style
 
 * style: prevent pandas warning ([`058a711`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/058a711a36998d1fb0bc44ee198615c8f660b1fe))
-
-### Unknown
-
-* 0.2.6
-
-Automatically generated by python-semantic-release
-
-[skip actions] ([`fc941a5`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/fc941a5f86c1a1b39ef84c80d88b72b6270fcdac))
-
-* Merge remote-tracking branch &#39;origin/main&#39; ([`e16069e`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/e16069e1ff2f5b3cda5b53fd1f75827f2ed6454c))
 
 
 ## v0.2.5 (2024-05-13)
@@ -151,15 +113,6 @@ Automatically generated by python-semantic-release
 
[…]
