### 2025-02-06 00aba82
docs: added contributing guide to docs
