### 2025-02-06 7698588
docs: moved examples to subpage
