+++ documentation/examples.rst
@@ -1,0 +1,12 @@
+
+Examples | Tutorial
+===================
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Contents:
+
+   examples/example_commandline
+   examples/example_python_api
+   examples/example_heavy_rainfall_index
+   examples/example_python_api_extended
