### 2025-02-06 f125432
style: minor code change in example
