### 2025-02-07 4bbc044
fix: fixed pandas warning when running function idf.add_max_return_periods_to_events(events)
