### 2025-02-07 d0a0852
docs: added some docstrings
