### 2025-02-07 fa5d430
docs: added some docstrings
