### 2025-02-07 fb709c7
style: removed redundant init function in design rainfall classes
