### 2025-02-07 086b3bc
docs: added some docstrings
