### 2025-02-07 ab3aec6
chore: added automatic tests with github automatisation
