### 2025-02-07 e6d2be1
style: clean up class code. group by kind of function
