+++ idf_analysis/idf_class.py
@@ -13,10 +13,11 @@ from .arg_parser import heavy_rain_parser
 from .definitions import *
 from .idf_backend import IdfParameters
 from .in_out import import_series
-from .little_helpers import minutes_readable, height2rate, delta2min, rate2height, frame_looper, event_caption, \
-    event_caption_ger, duration_steps_readable, get_progress_bar
+from .little_helpers import (minutes_readable, height2rate, delta2min, rate2height, frame_looper, event_caption,
+                             event_caption_ger, duration_steps_readable, get_progress_bar)
 from .plot_helpers import idf_bar_axes
-from .sww_utils import guess_freq, rain_events, agg_events, event_duration, resample_rain_series, rain_bar_plot, IdfError
+from .sww_utils import (guess_freq, rain_events, agg_events, event_duration, resample_rain_series, rain_bar_plot,
+                        IdfError)
 from .synthetic_rainseries import _BlockRain, _EulerRain
 
 
@@ -38,7 +39,8 @@ class IntensityDurationFrequencyAnalyse:
         _rainfall_sum_frame (pandas.DataFrame): with rain sums of all given durations
 
     """
-    def __init__(self, series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=False):
+
+    def __init__(self, series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=False, unit='mm'):
         """
         Heavy rainfall intensity as a function of duration and return period acc. to DWA-A 531 (2012).
 
@@ -51,6 +53,7 @@ class IntensityDurationFrequencyAnalyse:
             series_kind (str): ['partial', 'annual']
             worksheet (str): ['DWA-A_531', 'ATV-A_121', 'DWA-A_531_advektiv']
             extended_durations (bool): add [720, 1080, 1440, 2880, 4320, 5760, 7200, 8640] minutes to the calculation
+            unit (str, optional): Unit for rainfall (default: 'mm').
         """
         self._series = None
         self._freq = None
@@ -58,6 +61,8 @@ class IntensityDurationFrequencyAnalyse:
         #  how to calculate the idf curves
         self._parameters = IdfParameters(series_kind=series_kind, worksheet=worksheet,
                                          extended_durations=extended_durations)
+
+        self._unit = unit
 
         self._return_periods_frame = None
         self._rain_events = None
@@ -76,12 +81,13 @@ class IntensityDurationFrequencyAnalyse:
     def series(self, series: pd.Series):
         self._series = series
 
-    def set_series(self, series):
+    def set_series(self, series, unit='mm'):
         """
         set the series for the analysis
 
         Args:
             series (pandas.Series): precipitation time-series
+            unit (str, optional): Unit for rainfall (default: 'mm').
         """
         if not isinstance(series, pd.Series):
             raise IdfError('The series has to be a pandas Series.')
@@ -99,6 +105,7 @@ class IntensityDurationFrequencyAnalyse:
         self._return_periods_frame = None
         self._rain_events = None
         self._rainfall_sum_frame = None
+        self._unit = unit
 
     # __________________________________________________________________________________________________________________
     @property
@@ -184,7 +191,8 @@ class IntensityDurationFrequencyAnalyse:
         Extract interim results from parameters.
 
         Args:
-            filename (str, Path): filename of the parameters yaml-file
+            filename (str, Path): filename of the parameters yaml-file.
+            worksheet (str): ['DWA-A_531', 'ATV-A_121', 'DWA-A_531_advektiv']
         """
         self._parameters = IdfParameters.from_yaml(filename, worksheet)
 
@@ -202,14 +210,14 @@ class IntensityDurationFrequencyAnalyse:
[…]
