### 2025-02-07 f4825a2
fix: added language option in some backend functions
