+++ idf_analysis/little_helpers.py
@@ -69,6 +69,7 @@ def duration_steps_readable(durations):
 def height2rate(height_of_rainfall, duration):
     """
     calculate the specific rain flow rate in [l/(s*ha)]
+
     if 2 array-like parameters are give, an element-wise calculation will be made.
     So the length of the array must be the same.
 
@@ -85,6 +86,7 @@ def height2rate(height_of_rainfall, duration):
 def rate2height(rain_flow_rate, duration):
     """
     convert the rain flow rate to the height of rainfall in [mm]
+
     if 2 array-like parameters are give, an element-wise calculation will be made.
     So the length of the array must be the same.
 
@@ -113,7 +115,7 @@ def frame_looper(size, columns, label='return periods'):
         return columns
 
 
-def event_caption(event, unit='mm'):
+def event_caption_eng(event, unit='mm'):
     """
     Generates a human-readable caption for a rain event.
 
@@ -150,31 +152,7 @@ def event_caption(event, unit='mm'):
              The maximum return period was 10 years
              at a duration of 30 minutes."
     """
-    caption = 'rain event\n'
-    if (COL.START in event) and (COL.END in event):
-        caption += f'between {event[COL.START]:%Y-%m-%d %H:%M} and '
-        if f'{event[COL.START]:%Y-%m-%d}' == f'{event[COL.END]:%Y-%m-%d}':
-            caption += f'{event[COL.END]:%H:%M}\n'
-        elif f'{event[COL.START]:%Y-%m-}' == f'{event[COL.END]:%Y-%m-}':
-            caption += f'{event[COL.END]:%d %H:%M}\n'
-        else:
-            caption += f'{event[COL.END]:%Y-%m-%d %H:%M}\n'
-
-    if COL.LP in event:
-        caption += f'with a total sum of {event[COL.LP]:0.1f} {unit}\n'
-
-    if COL.DUR in event:
-        caption += f' and a duration of {timedelta_readable(event[COL.DUR])}'
-
-    caption += '.\n'
-
-    if COL.MAX_PERIOD in event:
-        caption += f' The maximum return period was {return_period_formatter(event[COL.MAX_PERIOD])} a\n'
-
-        if COL.MAX_PERIOD_DURATION in event:
-            caption += f' at a duration of {minutes_readable(event[COL.MAX_PERIOD_DURATION])}.'
-
-    return caption
+    return event_caption(event, unit=unit, lang='en')
 
 
 def event_caption_ger(event, unit='mm'):
@@ -214,44 +192,115 @@ def event_caption_ger(event, unit='mm'):
              Die maximale Wiederkehrperiode war 10 a
              bei einer Dauerstufe von 30 minutes."
     """
-    caption = 'Regenereignis'
+    return event_caption(event, unit=unit, lang='de')
+
+
+def event_caption(event, unit='mm', lang='en'):
+    """
+    Generates a human-readable caption for a rain event in English or German.
+
+    Args:
+        event (dict or pandas.Series): A dictionary or Series containing event details. Expected keys include:
+            - COL.START: Start time of the event.
+            - COL.END: End time of the event.
+            - COL.LP: Total rainfall sum (optional).
+            - COL.DUR: Duration of the event (optional).
+            - COL.MAX_PERIOD: Maximum return period (optional).
+            - COL.MAX_PERIOD_DURATION: Duration of the maximum return period in minutes (optional).
+        unit (str, optional): Unit for rainfall (default: 'mm').
+        lang (str, optional): Language ('en' for English, 'de' for German, default: 'en').
+
+    Returns:
+        str: A formatted string describing the rain event.
+
+    Example:
+        Given an event with:
[…]
