### 2025-02-07 e6e2839
test: added test for little helpers functions
