+++ tests/test_little_helpers.py
@@ -1,0 +1,267 @@
+import pytest
+import pandas as pd
+import datetime
+from datetime import timedelta
+
+from idf_analysis.definitions import COL
+from idf_analysis.little_helpers import (
+    delta2min, minutes_readable, duration_steps_readable,
+    height2rate, rate2height, timedelta_readable, timedelta_readable2, return_period_formatter,
+    event_caption)
+
+
+# Parameterized test for delta2min
+@pytest.mark.parametrize("input_value, expected", [
+    (pd.Timedelta(minutes=30), 30),
+    (pd.DateOffset(minutes=45), 45),
+    (pd.Timedelta(hours=1), 60),
+    (pd.DateOffset(hours=2), 120)
+])
+def test_delta2min(input_value, expected):
+    assert delta2min(input_value) == expected
+
+
+# Parameterized test for minutes_readable
+@pytest.mark.parametrize("input_value, expected", [
+    (30, "30 min"),
+    (120, "2 h"),
+    (1440, "1 d"),
+    (1500, "1.0 d"),
+    (2880, "2 d"),
+    (43200, "30 d")
+])
+def test_minutes_readable(input_value, expected):
+    assert minutes_readable(input_value) == expected
+
+
+def test_duration_steps_readable():
+    assert duration_steps_readable([30, 120, 1440]) == ["30 min", "2 h", "1 d"]
+
+
+# Parameterized test for height2rate with scalar values
+@pytest.mark.parametrize("height, duration, expected", [
+    (10, 60, 10 / 60 * (1000 / 6)),
+    (20, 120, 20 / 120 * (1000 / 6)),
+    (0, 60, 0.0)
+])
+def test_height2rate_scalar(height, duration, expected):
+    result = height2rate(height, duration)
+    assert isinstance(result, float)  # Check type
+    assert result == pytest.approx(expected, rel=1e-5)
+
+
+# Parameterized test for height2rate with pandas Series
+@pytest.mark.parametrize("heights, duration, expected", [
+    (pd.Series([10, 20, 30]), 60, pd.Series([10, 20, 30]) / 60 * (1000 / 6)),
+    (pd.Series([5, 15, 25]), 120, pd.Series([5, 15, 25]) / 120 * (1000 / 6))
+])
+def test_height2rate_series(heights, duration, expected):
+    result = height2rate(heights, duration)
+    assert isinstance(result, pd.Series)  # Check type
+    pd.testing.assert_series_equal(result, expected, atol=1e-5)  # Compare full series with tolerance
+
+
+# Parameterized test for rate2height with scalar values
+@pytest.mark.parametrize("rate, duration, expected", [
+    (2.77778, 60, 2.77778 * 60 / (1000 / 6)),
+    (4, 120, 4 * 120 / (1000 / 6)),
+    (0, 60, 0.0)
+])
+def test_rate2height_scalar(rate, duration, expected):
+    result = rate2height(rate, duration)
+    assert isinstance(result, float)  # Check type
+    assert result == pytest.approx(expected, rel=1e-5)
+
+
+# Parameterized test for rate2height with pandas Series
+@pytest.mark.parametrize("rate, duration, expected", [
+    (pd.Series([2.77778, 5.55556]), 60, pd.Series([2.77778, 5.55556]) * 60 / (1000 / 6)),
+    (pd.Series([1, 3, 7]), 120, pd.Series([1, 3, 7]) * 120 / (1000 / 6))
+])
+def test_rate2height_series(rate, duration, expected):
+    result = rate2height(rate, duration)
+    assert isinstance(result, pd.Series)  # Check type
+    pd.testing.assert_series_equal(result, expected, atol=1e-5)  # Compare full series with tolerance
+
+
[…]
