### 2025-02-07 408ebc4
test: added test for sww_utils functions
