### 2025-02-07 e9dc66c
test: added shell script to start tests
