### 2025-02-07 8adafec
test: added test for idf_backend functions
