+++ tests/test_idf_backend.py
@@ -1,0 +1,115 @@
+import pytest
+import numpy as np
+from idf_analysis.idf_backend import IdfParameters
+from idf_analysis.definitions import SERIES, METHOD, PARAM, APPROACH
+from idf_analysis.parameter_formulas import HyperbolicAuto, DoubleLogNormAuto, LinearFormula
+
+
+@pytest.mark.parametrize("extended, expected_durations", [
+    (False, np.array([5, 10, 15, 20, 30, 45, 60, 90, 120, 180, 240, 360, 540, 720, 1080])),
+    (True, np.array([5, 10, 15, 20, 30, 45, 60, 90, 120, 180, 240, 360, 540, 720, 1080,
+                     1440, 2880, 4320, 5760, 7200, 8640]))
+])
+def test_idf_durations(extended, expected_durations):
+    idf_params = IdfParameters(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=extended)
+    np.testing.assert_array_equal(idf_params.durations, expected_durations)
+
+
+@pytest.mark.parametrize("worksheet, expected_params", [
+    (METHOD.KOSTRA, {0: {'u': {'function': 'hyperbolic'}, 'w': {'function': 'double_logarithmic'}},
+                     60: {'u': {'function': 'double_logarithmic'}, 'w': {'function': 'double_logarithmic'}},
+                     720: {'u': {'function': 'linear'}, 'w': {'function': 'linear'}}}),
+    (METHOD.DWA_adv, {0: {'u': {'function': 'hyperbolic'}, 'w': {'function': 'double_logarithmic'}},
+                      180: {'u': {'function': 'double_logarithmic'}, 'w': {'function': 'double_logarithmic'}},
+                      1440: {'u': {'function': 'linear'}, 'w': {'function': 'linear'}}})
+])
+def test_idf_parameters_final(worksheet, expected_params):
+    idf_params = IdfParameters(series_kind=SERIES.PARTIAL, worksheet=worksheet, extended_durations=True)
+    assert idf_params.parameters_final == expected_params
+
+
+@pytest.mark.parametrize("filter_value, expected_durations", [
+    (15, np.array([15, 30, 45, 60, 90, 120, 180, 240, 360, 540, 720, 1080, 1440, 2880, 4320, 5760, 7200, 8640])),
+    (120, np.array([120, 240, 360, 720, 1080, 1440, 2880, 4320, 5760, 7200, 8640]))
+])
+def test_filter_durations(filter_value, expected_durations):
+    idf_params = IdfParameters(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=True)
+    idf_params.filter_durations(filter_value)
+    np.testing.assert_array_equal(idf_params.durations, expected_durations)
+
+
+def test_clear_and_add_parameter_approach():
+    idf_params = IdfParameters(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=False)
+    idf_params.clear_parameter_approaches()
+    assert idf_params.parameters_final == {}
+
+    idf_params.add_parameter_approach(60, APPROACH.LIN, APPROACH.LOG1)
+    expected = {60: {'u': {'function': 'linear'}, 'w': {'function': 'single_logarithmic'}}}
+    assert idf_params.parameters_final == expected
+
+
+@pytest.mark.parametrize("duration, param, expected", [
+    (30, PARAM.U, {'function': 'hyperbolic'}),
+    (300, PARAM.W, {'function': 'double_logarithmic'})
+])
+def test_get_duration_section(duration, param, expected):
[…]
