### 2025-02-07 dd896cb
test: added test for idf_class functions
