+++ tests/test_idf_class.py
@@ -1,0 +1,340 @@
+import os
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from idf_analysis import IntensityDurationFrequencyAnalyse
+from idf_analysis.definitions import SERIES, METHOD
+from idf_analysis.parameter_formulas import LinearFormula
+
+
+@pytest.fixture
+def idf():
+    """Fixture to initialize an IDF analysis instance and read parameters from a YAML file."""
+    idf = IntensityDurationFrequencyAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA,
+                                            extended_durations=False)
+    idf.read_parameters('../examples/ehyd_112086_idf_data/idf_parameters.yaml')
+    return idf
+
+
+@pytest.fixture
+def idf_r():
+    """Fixture to initialize an IDF analysis instance and read parameters from a idf table file."""
+    idf_table = pd.read_csv('../examples/ehyd_112086_idf_data/idf_table_UNIX.csv', header=[0, 1], index_col=0)
+    idf_table.columns = idf_table.columns.get_level_values(0).astype(int)
+    idf = IntensityDurationFrequencyAnalyse.from_idf_table(idf_table, linear_interpolation=True)
+    return idf
+
+
+@pytest.fixture
+def rainfall_series():
+    """Fixture to create a sample rainfall time series with 5-minute intervals and two rainfall events."""
+    freq_min = 5
+    start_time = pd.Timestamp("2024-01-01 00:00:00")
+    end_time = start_time + pd.Timedelta(hours=10)
+    time_index = pd.date_range(start=start_time, end=end_time, freq=f"{freq_min}min")
+
+    rainfall = pd.Series(0., index=time_index)
+
+    events = [
+        {"start": start_time + pd.Timedelta(minutes=10), "duration": 20, "intensity": 20.},
+        {"start": start_time + pd.Timedelta(hours=5, minutes=10), "duration": 60, "intensity": 50.}
+    ]
+
+    for event in events:
+        end_time = event["start"] + pd.Timedelta(minutes=event["duration"])
+        mask = (rainfall.index >= event["start"]) & (rainfall.index < end_time)
+        rainfall.loc[mask] = event["intensity"] / (event["duration"] / freq_min)
+    return rainfall
+
+
+@pytest.mark.parametrize("years, duration, expected", [
+    (2, 60,
+     [0, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903,
+      3.144903]),
+])
+def test_model_rain_block(idf, years, duration, expected):
+    result = idf.model_rain_block.get_series(years, duration)
+    np.testing.assert_allclose(result.values, expected, atol=1e-5)
+    expected_index = pd.Index(list(range(0, duration + 5, 5)))
+    pd.testing.assert_index_equal(result.index, expected_index)
+
+
+@pytest.mark.parametrize("years, duration, kind, expected", [
+    (2, 60, 2,
+     [0, 3.346922, 4.494913, 6.477299, 10.991037, 2.616501, 2.120819, 1.767846, 1.506858, 1.307946, 1.152510, 1.028474,
+      0.927712]),
+])
+def test_model_rain_euler(idf, years, duration, kind, expected):
+    result = idf.model_rain_euler.get_series(years, duration, kind=kind)
+    np.testing.assert_allclose(result.values, expected, atol=1e-5)
+    expected_index = pd.Index(list(range(0, duration + 5, 5)))
+    pd.testing.assert_index_equal(result.index, expected_index)
+
+
[…]
