### 2025-02-07 f9e9f98
fixed: fixed error when duration argument of function was not used.
