+++ CHANGELOG.md
@@ -1,6 +1,122 @@
 # CHANGELOG
 
-<!--next-version-placeholder-->
+
+
+## v0.3.1 (2025-02-07)
+
+### Chore
+
+* chore: added automatic tests with github automatisation ([`ab3aec6`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/ab3aec63a3f5863858abe3348d06b9cbe3e23da9))
+
+### Documentation
+
+* docs: fixed typo ([`c6169b9`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/c6169b9670f04b231110c3e9ea863d58e5b4ff3b))
+
+* docs: fixed a grammer error ([`2752694`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/27526941b1ecdb36d73e06a838f5facfaf7940a1))
+
+* docs: added some docstrings ([`086b3bc`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/086b3bc2fec580caaa66ee7cbb3e5940e2cb0b6f))
+
+* docs: added some docstrings ([`fa5d430`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/fa5d430cdd67e65d3ca2d018ad944af48a219931))
+
+* docs: added some docstrings ([`d0a0852`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/d0a085298d31539f644f807845ff32abad541f6e))
+
+* docs: added timeseries data format info to tutorial ([`f5f6b9e`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/f5f6b9e25cc6d75e7fe8b2de1b797ddcfb72c5c1))
+
+* docs: moved examples to subpage ([`7698588`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/769858891981dcb04fd524bc88d059a04e449bfc))
+
+* docs: added contributing guide to docs ([`00aba82`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/00aba8282f61a2da9b1e8e2a1ed341268ba40a95))
+
+* docs: clean up changelog ([`a096039`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/a0960393ecea3223e44aead7fe2ee5a16be63c04))
+
+* docs: clean up changelog ([`71ace68`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/71ace684972133a2d7e7e35b72a1ab251b8701b7))
+
+* docs: added contributing guide ([`1ca4904`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/1ca490439de02f446c91d0ffb858f70f4787e54b))
+
+* docs: updated readme ([`1298fd3`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/1298fd3743bd82b46f72a11076a54fc6fb7230f7))
+
+* docs: readme update ([`c509daa`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/c509daa74b4d4dcea1fbf61c25653820f4d20c92))
+
+* docs: update readme ([`a57edfc`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/a57edfc63bb258b0d9e3052f3b640228cc850d96))
+
+* docs: update readme ([`926a6a8`](https://github.com/MarkusPic/intensity_duration_frequency_analysis/commit/926a6a81483ea79bca55835da73e2d685ebe0897))
+
[…]
