### 2025-02-07 f25959e
chore: fixed automatic tests with github automatisation
