### 2025-02-07 0a69eda
chore: fixed automatic tests with github automatisation
