### 2025-02-07 d5444b1
chore: fixed automatic tests with github automatisation
