+++ .github/workflows/tests.yaml
@@ -19,7 +19,7 @@ jobs:
       - name: Install dependencies
         run: |
           pip install -r requirements.txt
-          pip install pytest pytest-cov
+          pip install pytest pytest-cov pyarrow
           pip install -e .
 
       - name: Run tests with coverage
