### 2025-02-07 ea26c37
chore: fixed automatic tests with github automatisation
