### 2025-02-07 6ddd636
chore: fixed automatic tests with github automatisation
