+++ .github/workflows/tests.yaml
@@ -29,7 +29,9 @@ jobs:
         run: ls -l coverage.xml || { echo "coverage.xml not found!"; exit 1; }
 
       - name: Upload coverage report
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v5
         with:
-          file: ./coverage.xml
+          token: ${{ secrets.CODECOV_TOKEN }}
+          slug: MarkusPic/intensity_duration_frequency_analysis
+          #file: ./coverage.xml
           fail_ci_if_error: true
