### 2025-02-07 93894e7
chore: simplify script for automatic tests with github automatisation
