### 2025-02-07 8505916
chore: added testing dependencies to pyproject.toml
