+++ pyproject.toml
@@ -35,6 +35,7 @@ dynamic = ["version"]
 
 [project.optional-dependencies]
 docs = ["sphinx", "nbsphinx", "recommonmark", "myst-parser", "pydata_sphinx_theme", "ipython==8.5.0", 'sphinx-codeautolink']
+testing = ["pytest", "pytest-cov", "pyarrow"]
 
 [project.urls]
 Documentation = "https://markuspic.github.io/intensity_duration_frequency_analysis/"
