### 2025-02-07 512ab11
style: simplify function minutes_readable
