### 2025-02-07 e77c4a8
style: remove unused internal function
