+++ idf_analysis/little_helpers.py
@@ -113,86 +113,6 @@ def frame_looper(size, columns, label='return periods'):
         return columns
 
 
-def event_caption_eng(event, unit='mm'):
-    """
-    Generates a human-readable caption for a rain event.
-
-    The caption includes details such as the event's start and end times, total rainfall, duration,
-    and maximum return period (if available). The output is formatted for readability.
-
-    Args:
-        event (dict or pandas.Series): A dictionary or Series containing event details. Expected keys include:
-            - COL.START: Start time of the event.
-            - COL.END: End time of the event.
-            - COL.LP: Total rainfall sum (optional).
-            - COL.DUR: Duration of the event (optional).
-            - COL.MAX_PERIOD: Maximum return period (optional).
-            - COL.MAX_PERIOD_DURATION: Duration of the maximum return period (optional).
-        unit (str, optional): Unit for rainfall (default: 'mm').
-
-    Returns:
-        str: A formatted string describing the rain event.
-
-    Example:
-        Given an event with:
-            - COL.START = pd.Timestamp('2023-01-01 12:00')
-            - COL.END = pd.Timestamp('2023-01-01 14:00')
-            - COL.LP = 15.5
-            - COL.DUR = pd.Timedelta(hours=2)
-            - COL.MAX_PERIOD = 10
-            - COL.MAX_PERIOD_DURATION = pd.Timedelta(minutes=30)
-
-        The output might look like:
-            "rain event
-             between 2023-01-01 12:00 and 14:00
-             with a total sum of 15.5 mm
-             and a duration of 2 hours.
-             The maximum return period was 10 years
-             at a duration of 30 minutes."
-    """
-    return event_caption(event, unit=unit, lang='en')
-
-
-def event_caption_ger(event, unit='mm'):
-    """
-    Generates a human-readable caption for a rain event in german.
-
-    The caption includes details such as the event's start and end times, total rainfall, duration,
-    and maximum return period (if available). The output is formatted for readability.
-
-    Args:
-        event (dict or pandas.Series): A dictionary or Series containing event details. Expected keys include:
-            - COL.START: Start time of the event.
-            - COL.END: End time of the event.
-            - COL.LP: Total rainfall sum (optional).
-            - COL.DUR: Duration of the event (optional).
-            - COL.MAX_PERIOD: Maximum return period (optional).
-            - COL.MAX_PERIOD_DURATION: Duration of the maximum return period (optional).
-        unit (str, optional): Unit for rainfall (default: 'mm').
-
-    Returns:
-        str: A formatted string describing the rain event.
-
-    Example:
-        Given an event with:
-            - COL.START = pd.Timestamp('2023-01-01 12:00')
-            - COL.END = pd.Timestamp('2023-01-01 14:00')
-            - COL.LP = 15.5
-            - COL.DUR = pd.Timedelta(hours=2)
-            - COL.MAX_PERIOD = 10
-            - COL.MAX_PERIOD_DURATION = pd.Timedelta(minutes=30)
-
-        The output might look like:
-            "Regenereignis
-             von 01.01.2023 12:00 bis 14:00
-             mit einer Regensumme von 15.5 mm
-             und einer Dauer von 2 Stunden.
-             Die maximale Wiederkehrperiode war 10 a
-             bei einer Dauerstufe von 30 minutes."
-    """
-    return event_caption(event, unit=unit, lang='de')
-
-
 def event_caption(event, unit='mm', lang='en'):
     """
     Generates a human-readable caption for a rain event in English or German.
