### 2025-02-07 e6f7c98
style: simplified internal functions
