### 2025-02-07 4d74f8f
style: simplified internal functions
