### 2025-02-07 7f73c22
test: added test for sri_class functions
