+++ tests/test_sri_class.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from idf_analysis import IntensityDurationFrequencyAnalyse
+from idf_analysis import HeavyRainfallIndexAnalyse
 from idf_analysis.definitions import SERIES, METHOD
 from idf_analysis.parameter_formulas import LinearFormula
 
@@ -16,8 +16,8 @@ PTH_EXAMPLES = Path(__file__).parent.parent / "examples"
 @pytest.fixture
 def idf():
     """Fixture to initialize an IDF analysis instance and read parameters from a YAML file."""
-    idf = IntensityDurationFrequencyAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA,
-                                            extended_durations=False)
+    idf = HeavyRainfallIndexAnalyse(series_kind=SERIES.PARTIAL, worksheet=METHOD.KOSTRA, extended_durations=False,
+                                    method=HeavyRainfallIndexAnalyse.METHODS.SCHMITT)
     idf.read_parameters(PTH_EXAMPLES / 'ehyd_112086_idf_data/idf_parameters.yaml')
     return idf
 
@@ -27,7 +27,7 @@ def idf_r():
     """Fixture to initialize an IDF analysis instance and read parameters from a idf table file."""
     idf_table = pd.read_csv(PTH_EXAMPLES / 'ehyd_112086_idf_data/idf_table_UNIX.csv', header=[0, 1], index_col=0)
     idf_table.columns = idf_table.columns.get_level_values(0).astype(int)
-    idf = IntensityDurationFrequencyAnalyse.from_idf_table(idf_table, linear_interpolation=True)
+    idf = HeavyRainfallIndexAnalyse.from_idf_table(idf_table, linear_interpolation=True)
     return idf
 
 
@@ -53,102 +53,31 @@ def rainfall_series():
     return rainfall
 
 
-@pytest.mark.parametrize("years, duration, expected", [
-    (2, 60,
-     [0, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903, 3.144903,
-      3.144903]),
-])
-def test_model_rain_block(idf, years, duration, expected):
-    result = idf.model_rain_block.get_series(years, duration)
-    np.testing.assert_allclose(result.values, expected, atol=1e-5)
-    expected_index = pd.Index(list(range(0, duration + 5, 5)))
-    pd.testing.assert_index_equal(result.index, expected_index)
-
-
-@pytest.mark.parametrize("years, duration, kind, expected", [
-    (2, 60, 1, [0., 10.991037, 6.477299, 4.494913, 3.346922, 2.616501, 2.120819, 1.767846, 1.506858, 1.307946, 1.15251, 1.028474, 0.927712]),
-    (2, 60, 2, [0, 3.346922, 4.494913, 6.477299, 10.991037, 2.616501, 2.120819, 1.767846, 1.506858, 1.307946, 1.152510, 1.028474, 0.927712]),
-    (2, 75, 2, [0., 2.616501, 3.346922, 4.494913, 6.477299, 10.991037, 2.120819, 1.767846, 1.506858, 1.307946, 1.15251, 1.028474, 0.927712, 0.522366, 0.490181, 0.46208]),
-    (2, 75, 3, [0., 2.120819, 2.616501, 3.346922, 4.494913, 6.477299, 10.991037, 1.767846, 1.506858, 1.307946, 1.15251, 1.028474, 0.927712, 0.522366, 0.490181, 0.46208]),
-])
-def test_model_rain_euler(idf, years, duration, kind, expected):
[…]
