### 2025-02-07 e0ed961
test: added test for sri_class functions
