### 2025-02-07 788e556
test: added test for arg_parser and plot_helpers functions
