### 2025-02-10 #7607 @MarkusPic
@RBhupi, @einaraz, @cheginit I made some updates:

### Software Enhancements
- **Tutorial Update**: Added a detailed paragraph explaining the expected format of timeseries data, including required units.
- **Contributing Guide**: Introduced a `CONTRIBUTING.md` file to provide clear guidelines for contributions and support.
- **Testing Improvements**: Implemented automated unit tests and added a coverage report to ensure robust code quality.

### Paper Revisions (will be updated today)
- **Figure Clarity**: Enhanced the figure caption by including descriptions of the x-axis, y-axis, and legend for better interpretation.
- **Equation Context**: Included a clear definition of *D* below the equation for better understanding.
- **Content Refinement**:
  - Rephrased the section on partial series for improved clarity.
  - Rewrote the paragraph discussing previous literature and tools to enhance flow and accuracy.
