+++ joss-paper/paper.bib
@@ -22,7 +22,7 @@
   doi = {10.5281/zenodo.13896180},
 }
 
-@article{Lanciotti:2022,
+@article{lanciotti:2022,
   author = {Lanciotti, Sabrina and Ridolfi, Elena and Russo, Fabio and Napolitano, Francesco},
   title = {Intensity–Duration–Frequency Curves in a Data-Rich Era: A Review},
   journal = {Water},
@@ -34,7 +34,7 @@
   doi = {10.3390/w14223705}
 }
 
-@article{Schardong:2020,
+@article{schardong:2020,
   author = {Schardong, Andre and Simonovic, Slobodan P. and Gaur, Abhishek and Sandink, Dan},
   title = {Web-Based Tool for the Development of Intensity Duration Frequency Curves under Changing Climate at Gauged and Ungauged Locations},
   journal = {Water},
@@ -46,11 +46,25 @@
   doi = {10.3390/w12051243}
 }
 
-@software{Mendez:2024,
+@software{mendez:2024,
   author = {Diego Armando Urrea Méndez},
   title = {IDF},
   year = {2024},
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/diegourreamendez/IDF},
+}
+
+
+@article{cunnane:1978,
+title = {Unbiased plotting positions — A review},
+journal = {Journal of Hydrology},
+volume = {37},
+number = {3},
+pages = {205-222},
+year = {1978},
+issn = {0022-1694},
+doi = {https://doi.org/10.1016/0022-1694(78)90017-3},
+url = {https://www.sciencedirect.com/science/article/pii/0022169478900173},
+author = {C. Cunnane},
 }
