### 2025-02-11 9e13626
docs: joss paper fix minor typo
