### 2025-02-14 #7607 @einaraz
Hi @MarkusPic. Thank you for addressing my comments in the revised version of the package and paper. The new version is significantly clearer and addresses my previous concerns. @cheginit , I believe it is now ready for publication.
Best,
Einara
