### 2025-02-18 #7607 @cheginit
@MarkusPic please address the following:

- [ ] Fix the DOI issues found by the editorialbot
- [ ] L11-14: This paragraph is a bit difficult to read. Consider combining the two similar statements about accuracy into a single clear statement about how accuracy affects both costs and risks.
- [ ] L16-17: The pronoun "them" could be clearer - specify whether it refers to "values" or "methods and models".
- [ ] L26-31: This informative sentence would flow better if split into two parts. Consider breaking after "IDF curves" and starting a new sentence with "Additionally".
- [ ] L34-35: The flow would improve by restructuring to: "When analyzing long-term datasets spanning at least 20 years (ideally over 50 years)..."
- [ ] L36-39: Consider splitting this sentence into two parts at "The number of selected events..." for better readability.
- [ ] L53-56: A small restructure could help: "While Méndez (2024) introduced a similar tool, it has limitations: it only supports..."
- [ ] L61-67: The ideas in this paragraph would be clearer if the longer sentences were broken into smaller parts, especially the sentence spanning lines 62-64.
- [ ] L72-74: The acknowledgment to David Chamy could flow more smoothly as a single combined sentence rather than across multiple lines.
