### 2025-02-18 741bb4b
docs: joss paper update after review of the editor.
