+++ joss-paper/paper.md
@@ -23,9 +23,7 @@ bibliography: paper.bib
 
 Heavy rainfall data are among the most important planning parameters in water management and hydraulic engineering practice. They are required as initial parameters for the dimensioning of rainwater drainage systems in urban areas and for the dimensioning of hydraulic structures on watercourses. 
 
-The accuracy of the target values of the corresponding calculation methods and models depends crucially on their accuracy. Their overestimation can lead to considerable additional costs in the structural implementation, their underestimation to an unacceptable, excessive residual risk of failure during the operation of water management and hydraulic engineering facilities. 
-
- It should be noted that overestimating them can lead to considerable additional costs during construction, while underestimating them can lead to an unacceptable, excessive residual risk of failure during the operation of water management and hydraulic engineering facilities.
+The accuracy of the calculation methods and models significantly impacts both costs and risks. Overestimation can lead to considerable additional construction costs, while underestimation can result in an excessive residual risk of failure during the operation of water management and hydraulic engineering facilities.
 
 An intensity-duration-frequency (IDF) curve is a mathematical function (see \autoref{eq:idf}) that relates the rainfall intensity with its duration and frequency of occurrence. An example can be seen in \autoref{fig:idf_curves}. These curves are commonly used in hydrology for flood forecasting and civil engineering for urban drainage design. However, the IDF curves are also analysed in hydrometeorology because of the interest in the time concentration or time-structure of the rainfall.
 
@@ -36,9 +34,9 @@ The package follows the guidance recommendations in @dwa:2012.
 
 # Statement of need
 
-`idf-analysis` is a Python package designed to analyze local changes in rainfall behavior and assess variations in return periods of specific rainfall events. It processes time-series rainfall measurements to derive parameters for intensity-duration-frequency (IDF) curves, enabling estimations of return periods for events of different durations. Additionally, the package allows for comparisons of rainfall behavior across different locations or time periods, making it useful for studying climate change effects.
+`idf-analysis` is a Python package designed to analyze local changes in rainfall behavior and assess variations in return periods of specific rainfall events. It processes time-series rainfall measurements to derive parameters for intensity-duration-frequency (IDF) curves. This enables estimations of return periods for events of different durations. Additionally, the package allows for comparisons of rainfall behavior across different locations or time periods. This makes it useful for studying climate change effects.
 
-The analysis begins by identifying independent rainfall events within the dataset. For each event, the maximum intensity is determined across a wide range of durations. When working with long-term datasets (ideally exceeding 50 years, but at least 20 years), an annual series analysis is performed using the Gumbel distribution, where only the maximum intensity per year is considered. For medium-length datasets (preferably over 20 years, but at least 10 years), a partial series is constructed by selecting the largest intensities based on an exponential distribution. The number of selected events is determined as 2.7 times the recording duration in years (e.g., for a 10-year dataset, the 27 highest intensities are analyzed).
+The analysis begins by identifying independent rainfall events within the dataset. For each event, the maximum intensity is determined across a wide range of durations. When analyzing long-term datasets spanning at least 20 years (ideally over 50 years), an annual series analysis is performed using the Gumbel distribution, where only the maximum intensity per year is considered. For medium-length datasets spanning at least 10 years (preferably over 20 years), a partial series is constructed by selecting the largest intensities based on an exponential distribution. The number of selected events is determined as 2.7 times the recording duration in years (e.g., for a 10-year dataset, the 27 highest intensities are analyzed).
 
[…]
