+++ joss-paper/paper.bib
@@ -64,7 +64,7 @@ number = {3},
 pages = {205-222},
 year = {1978},
 issn = {0022-1694},
-doi = {https://doi.org/10.1016/0022-1694(78)90017-3},
+doi = {10.1016/0022-1694(78)90017-3},
 url = {https://www.sciencedirect.com/science/article/pii/0022169478900173},
 author = {C. Cunnane},
 }
