### 2025-02-18 #7607 @MarkusPic
@RBhupi and @einaraz once again thank you both for the constructive review.

@cheginit
I worked through every point in the Post-Review Checklist
The latest release with all the changes suggested in the review is the version `0.3.1`
The DOI for the Zenodo archive is `10.5281/zenodo.14834297`

* [x] Fix the DOI issues found by the editorialbot
  - 🟡 SKIP DOIs
    - No DOI given, and none found for title: Starkregen in Abhängigkeit von Wiederkehrzeit und ...
      - there is no DOI to add
  - ❌ MISSING DOIs
    - 10.1007/springerreference_65082 may be a valid DOI for title: IDF
      - this DOI is wrong and there is no DOI
  - ❌ INVALID DOIs
      - I fixed this error in the bib file.
* [x] L11-14: This paragraph is a bit difficult to read. Consider combining the two similar statements about accuracy into a single clear statement about how accuracy affects both costs and risks.
* [x] L16-17: The pronoun "them" could be clearer - specify whether it refers to "values" or "methods and models".
* [x] L26-31: This informative sentence would flow better if split into two parts. Consider breaking after "IDF curves" and starting a new sentence with "Additionally".
  * I split two of the sentences in the paragraph to make it better readable.
* [x] L34-35: The flow would improve by restructuring to: "When analyzing long-term datasets spanning at least 20 years (ideally over 50 years)..."
* [ ] L36-39: Consider splitting this sentence into two parts at "The number of selected events..." for better readability.
  * A new sentence already starts at this point, so I don't understand
* [x] L53-56: A small restructure could help: "While Méndez (2024) introduced a similar tool, it has limitations: it only supports..."
* [x] L61-67: The ideas in this paragraph would be clearer if the longer sentences were broken into smaller parts, especially the sentence spanning lines 62-64.
* [x] L72-74: The acknowledgment to David Chamy could flow more smoothly as a single combined sentence rather than across multiple lines.
  * I rewrote the acknowledgments.
