### 2025-02-20 #7607 @cheginit
@MarkusPic Thanks for addressing all the issue. Please confirm whether v0.3 is the version that contain all the fixes that you made to address issue raised by the reviewers. If not, please release a new version and give me the version number. Then generate a DOI using an archiving service such as Zenodo, making sure that the title, version, author name and DOI, exactly match your submission.
