### 2025-02-20 #7607 @MarkusPic
@cheginit I'm sorry, maybe I misunderstood, but I commented on this at the beginning of my previous answer above.

- I worked through every point in the Post-Review Checklist
- The latest release with all the changes suggested in the review is the version `0.3.1`
- The DOI for the Zenodo archive is `10.5281/zenodo.14834297`

Or do I have to put this metadata somewhere else?
