### 2025-02-20 #7607 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
