### 2025-02-24 #7607 @kthyng
Congratulations on your new publication @MarkusPic! Many thanks to editor @cheginit and to reviewers @einaraz, @igarousi, and @RBhupi for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@MarkusPic I see you are in the JOSS reviewer database, which is awesome, but it would help if you add some language and topic area information (https://reviewers.joss.theoj.org/).
